Implement password storage for transport ssh_async
Because the password is something related to the computer it should be
associated also with the orm.Computer class. Also passing the password
just in auth parameters is unsafe as it is used for CLI suggestions.

Because we want to delete the entry in the secure storage one deletion
of Computer we need to put this implementation into BackendComputer. To
separate the interface of the computer properties that are stored
in the database we create a nested class ComputerPasswordManager and
BackendComputer has one instance of it.

To implement deletion with sqlalchemy we have to support the creation of
the PasswordManager with the DbComputer, because DbComputer is not
accesible as typehint we use a Protocol (implementation of HasUuid).

We also now use the uuid to store it in the database, this way we do not
have to consider problems with relabeling of computer.

Add tests for deletion of computer, verify that password is deleted

Add tests to test_all_plugins by adding the use cases parameter that
specifies if ssh_async should be tested with key or by password. I think
better design would be to separate the fixture custom_transport into
several ones and then pass it, I would prefer such a big refactor in a
different PR.