Add `_updatable_objects` to Node
All attributes in AiiDA are stored in the SQL database, which can become inefficient for handling **large data**. For example, the `checkpoint` data in WorkChain, or the `pickled` binary data in WorkGraph. To improve performance, it would be beneficial to store large data in the file repository (`base.repository`) instead.  However, unlike `attributes`, the repository does not support `updatable` objects in the same way as `updatable` attributes.  This PR introduces `_updatable_objects` to enable **mutable** repository storage for active processes while preserving AiiDA's immutability rules.