`CI`: `RTD` tail errors and warnings (#6776)
Previous to this commit, if RTD fails is a challenge to find what failed, because the tail of the log was not immediately descriptive:

```
(good) examples/wannier90_tutorial.html --> external_redirects.html
build finished with problems, 2 warnings.
```

With this commit, the tail in case of a failed build will look like this:

```
(good) examples/wannier90_tutorial.html --> external_redirects.html
build finished with problems, 2 warnings.
/aiida-core/src/aiida/calculations/stash.py:docstring of aiida.calculations.stash.StashCalculation:9: ERROR: Unexpected indentation. [docutils]
/aiida-core/src/aiida/calculations/stash.py:docstring of aiida.calculations.stash.StashCalculation:14: WARNING: Block quote ends without a blank line; unexpected unindent. [docutils]
```


Co-authored-by: Daniel Hollas danekhollas@gmail.com