Add force option to kill command
The force option gives the option to skip any action the process does to
properly release resources. This can result in orphan resources (e.g.
submitted jobs that are still running).

To test the feature we have to monkey patch the workers code. We thus
start a worker in a forked process to inherit the monkey patched
python context. The fixture `fork_worker_context` returns a
context manager that starts the worker with the aiida test profile and
terminates it.