Docs: Fix the logic in FizzBuzz WorkChain example
The logic of the FizzBuzz example is wrong and, as it stands, it would never trigger `FizzBuzz`.  Also, the increment needs to be _in_ the loop. This is a Programming 101 example that should probably be correct, especially since "intention of this example is to show that with a well designed outline, a user only has to look at the outline to have a good idea what the work chain does and how it does it." 

In addition, there were typos in the function names and the process should probably start from 1.