👌 Align behavior of `puttree` method for nested folders
Currently, trying to copy a nested folder such as:

```
dir
└── sub
    └── file
```

into a nested folder whose parent folder doesn't exist works perfectly fine for
`LocalTransport.puttree()`, since this command relies on `shutil.copytree` which allows
for this and reproduces the same file hierarchy. However, it fails for the
`SshTransport` and its async variant because the parent folder `dir` doesn't exist when
the `mkdir` command tries to create the `sub` directory.

Here we adapt the `puttree` command of both SSH transports to use the `makedirs` command
instead, ignoring the existence of the `sub` directory in case `overwrite` is set to
`True`.

Co-authored-by: superstar54 <xingwang1991@gmail.com>