🐛 Fix `local_copy_list` behavior for nested target folder
When using the `local_copy_list` to copy a nested folder from a `FolderData` node to a
a nested folder in the workdir, e.g.:

```
local_copy_list = [(uuid 'out/HP', 'out/HP'), ]
```

The resulting directory tree would be incorrect:

```
out
└── HP
    └── HP
        └── file
```

Even without any other copy operations (sandbox, remote) being executed. Here we adapt
the `put_async` command to copy the target directory instead of the whole temporary
directory contents in case of `FileType.DIRECTORY` to fix this issue.

We also verified that this behavior corresponds to the pre-v2.6 one, when the feature
to change the order of the copy lists was released.

Co-authored-by: superstar54 <xingwang1991@gmail.com>
Co-authored-by: Ali Khosravi <khsrali@gmail.com>