Add force-kill option when killing a process (#6793)
The force option gives the option to skip any action the process does to
properly release resources. This can result in orphan resources (e.g.
submitted jobs that are still running).

To test the feature we have to monkey patch the workers code. We thus start a
worker in a spawned process with a function that monkey patches the
corresponding modules. The spawned process has to do the monekypatching by
itself so we pass a function `prepare_func_args` to the spawned process when
initialized For correct usage, the fixture `fork_worker_context` should be used
that returns context manager starting the worker with the aiida test profile
and terminates it on exit.

Furthermore the `timeout` option of `verdi process kill` was not correctly
passed to the inner blocking `kiwipy.Future` resulting in a blocking call. This
has been fixed.

This change goes hand in hand with the change in PR aiidateam/plumpy/320.