Align the yes/no prompt in `verdi computer delete` with other prompts
The yes/no prompt of `verdi computer delete` is the only prompt in the CLI that
does not accept 'y' as valid answer but only 'yes'. This has been aligned with
the other prompts to also support 'y'.

Furthermore, no prompt is invoked when no nodes are associated with the
computer.

The tests have been restructured to more granular test cases and the names have
been cleaned up

We had to increase check count in `MockProcess` to avoid the timeout.