Fixing doc references after `click` 8.2 release (#6880)
`click` released v8.2.0 and by that broke the references to their doc. We pinned
`click~=8.1` so we still need to reference the old API but because `click` only
keeps one version of the documentation, we cannot reference a different API.
Thus we reference the functions by the GitHub link for v8.1.0.