StashCalculation: throw a warning if the computer node mismatches with the one of source_node (#6862)
We don't raise an error, but a warning. Because there could be a situation that a user's computer will be faulty for whatever reason (migrating transport plugins, etc.). Therefore it should still be possible to stash data from a different computer as long as they share access to the RemoteData path.