Include usage of `filter_size` for importing logs, user, computer
Building the query for the matching nodes from the backend can for
surpass the sqlite limit of allowed query parameters and thus result in
failure when importing the archive. In this PR we impement the usage of
the `filter_size` which limits the number of parameters in the query for
importing logs, user and computers.