Devops: Skip upload step in benchmarks if not pushed to main (#6895)
The if-condition that checks if the benchmark workflow should be run does not
consider when a branch from a fork is requested to be merged to the upstream
repo. In this case the workflow will run but will fail during the upload step
because the access to the secrets is not given. Since the upload of the results
is only needed for merged changes we only execute this step on a push event to
main.