Store default profile in `aiida_profile_factory` (#6893)
Ensures that the config set in fixture `aiida_profile_factory` is stored
persistently on disk and not just in memory. Tests have been added that use the
`aiida_config_tmp` implicitly through autouse `aiida_profile` fixture. The test
checks if the default profile is set when loading the config from file. Note that
`aiida_profile_factory` is intended to be used with a path argument to chose
the configuration directory. By that we avoid race conditions on configuration
tests due to concurrent tests.

The tests did not fail before because only the default profile option was not
stored persistently while the profile was, and further all CLI tests specify
the profile explicitly.