Sending multiple `process_kill` actions reschedules the cancelation of scheduler job (#6870)
PR #6793 introduced logic to cancel earlier kill actions when a new one is
issued. However, this led to a bug: when two kill actions are sent in
succession, the second cancels the first including its triggered cancelation of
the scheduler job that is stuck in the EBM. The second kill command does not
re-initiate the scheduler job cancelation, leaving it in an inconsistent state.
This issue arises because the kill logic is split across two places. PR #6868
addresses this properly with a refactor of the kill action. In contrast, this PR
provides a temporary workaround to mitigate the issue.

Furthermore, before this PR if the kill command failed through the EBM, the
cancelation could not be rescheduled by a subsequent kill action as the process
state already transition to the `Expected` state. With the new force-kill option
the user can bypasses the EBM as desired, thus the transition to the `Expected`
state is not needed anymore and has been removed to allow the user to further
gracefully kill a job.

---------

Co-authored-by: Ali Khosravi <khsrali@gmail.com>