Fix wrong config folder setup in `aiida_instance` fixture
This bug was introduced in PR #6610, where the temporary config folder
path was changed unintentionally. Previously, it was set to
`/tmp/pytest-.../config/.aiida`, but the new setup used
`/tmp/pytest-.../config/`. As a result, the daemon_client fixture failed
to detect the profile, leading to subtle issues.

This bug was masked by two factors:
 - The CI setup (`setup.sh`) which preconfigures AiiDA profiles.
 - The fixture definitions in `tests/conftest.py`, which interfered with
   the intended isolation of aiida_instance.

A follow-up commit will address this entanglement by clearly separating
fixture configuration and test logic.