Profile data dumping (#6723)
Data dumping for groups and profiles (#6723)

This PR adds functionality to incrementally dump profile and group data
into a human-readable output folder, and refactors the internal logic of
the recently released process dumping.

### Public API

The data dumping feature can be used from the CLI via the 
`verdi {profile|group|process} dump` commands. Furthermore, 
the classes `aiida.manage.configuration.Profile`, `aiida.orm.Group` and
`aiida.orm.ProcessNode` are extended by a new public member function
`dump` that takes the same dumping options as the CLI entry points.

The internal implementation of the feature is contained in the private
module `aiida.tools._dumping`, which is currently excluded from
`codecov`. Further testing and modifications will be applied based on
user feedback in smaller, more manageable PRs.

### Configuration of dumping

To organize the extensive options, a data class for the config options is
created using `pydantic` `BaseModel` in the `aiida.tools._dumping.config`
module. For each type of dumping (process/group/profile) different
options are available contained in three config classes
`ProcessDumpConfig`, `ProfileDumpConfig`, `GroupDumpConfig` all
inheriting from `BaseDumpConfig`. The `*DumpConfig` classes use the
mixin pattern to organize different options via the `TimeFilterMixin`,
`EntityFilterMixin`, `ProcessHandlingMixin` and `GroupManagementMixin`
since they are not available for each type of entity being dumped. The
new CLI entry points and the new member function `dump` all map their
inputs to the respective config class. By mapping both inputs to the
`*DumpConfig` classes the validation process is unified, reducing code
duplication.

### State of dumping folder

The dumping functionality tracks the state of the dumped folder wrt. to
the AiiDA database. This requires a persistent storage of the current
state of the dumping folder as well as logic comparing the state of the
dumping folder and the database. To prevent expensive file reads of the
dumping folder, the state is stored in a `json` file after the dumping
process. The logic to evaluate and track the state and compare it with
the database is contained in the modules `aiida.tools._dumping.tracking`
and `aiida.tools._dumping.mapping`, and changes since the last dump
(new/deleted nodes/groups, relabeled groups, node membership changes,
etc.) are then picked up via the module `aiida.tools._dumping.detector`.
This enables incremental dumping of data in a way that the
human-readable output folder of the dumping feature tracks the state of
the AiiDA DB as it evolves.

### Execution of the dump

The `aiida.tools._dumping.engine` module is responsible for the
top-level orchestration of the dumping process (including reading in the
`json` state file from the previous dump, or deleting it, if overwrite
mode is selected), as well as common setup and teardown operations. For
group and profile dumping, changes in AiiDA's DB since the last dump
that are not yet reflected in the dumping output folder, are then
carried out incrementally. This includes deleting output directories of
nodes and groups that were previously dumped but were since deleted from
AiiDA's DB, applying group relabeling carried out by the user, as well
as dumping new nodes and groups. This functionality is contained in the
`aiida.tools._dumping.executors` that provides the `DeletionExecutor`,
`ProcessDumpExecutor`, `GroupDumpExecutor`, and `ProfileDumpExecutor`,
classes and presents the meat of the feature implementation.

Finally, code that was previously contained in the `ProcessDumper` class
is now moved to the `ProcessDumpExecutor`, while a `ProcessDumper`
"facade" class is still provided via the `aiida.tools._dumping.facades`
module, and exposed as public API for backwards compatibility.