Deprecate in `verdi quicksetup` the `--db-engine` option (#6906)
The `--db-engine` option only supports `postgresql_psycopg`. To avoid failure
due the deprecation of `postgresql_psycopg2` (note that unintuitively
`postgresql_psycopg` is the newer version) we add the support in the click
options and echo a deprecation warning that the option does not have any effect.
This change does not have any effect on the actual DB engine being used as it is
hardcoded.

Fixes renaming of `database` to `dbname` missed in #6362 that raised
deprecation warning.