Include usage of `filter_size` for importing logs, user, computer (#6889)
Building the query for the matching nodes from the backend can for
surpass the sqlite limit of allowed query parameters and thus result in
failure when importing the archive. In this PR we implement the usage of
the `filter_size` which limits the number of parameters in the query for
importing logs, user and computers.

Fixes #6876 that is due to `_import_logs`.