Rename `machine_or_host` to `host` and simplify prompts
By only using the term `host` we do not need to consider both names
simplifying the prompt texts. Since the keyword we want the user to
specify is called `Host` in the ssh config, it should be clear what is
meant by this. While we also support the case if a password-less
connection to a machine can be achieved without specifying it in the ssh
config (e.g. by default ssh key), by ignoring this possibility it
simplifies the prompt texts. In addition, it is explained that the
`hostname` in the aiida computer setup is ignored which we still have
to continue support of `core.ssh`.