`Transport`: fixing the suggestion for parameter `host` (#6919)
Turns out the name of the suggestive function _get_<PARAMETER>_suggestion_string is not arbitrary.
And had to be adopted with a change of name in PR #6914 (commit  bf34c95)

```bash
$ verdi -p presto-3 computer configure core.ssh_async delete_me
Report: enter ? for help.
Report: enter ! to ignore the default and set no value.
Host as in 'ssh <HOST>' (needs to be a password-less setup in your ssh config): 
```

vs

```bash
$ verdi -p presto-3 computer configure core.ssh_async delete_me
Report: enter ? for help.
Report: enter ! to ignore the default and set no value.
Host as in 'ssh <HOST>' (needs to be a password-less setup in your ssh config) [HOST]: 
```