Move pytest fixture test to .github folder
The pytest fixture tests need to be isolated from the regular test suite
to test if they can work independently from aiida-core test suite.
Having them in the test suite risks them to hiddenly reuse
configurations as it introduced the bug #6904. Moving them to src means
however that we ship the tests with releases which we also want to
avoid. We therefore move them to the .github folder.