Move pytest fixture test to .github folder (#6921)
The pytest fixture tests need to be isolated from the regular test suite to
verify that they can function independently of the aiida-core test suite.
Including them in the main test suite risks inadvertently reusing
configurations, which previously led to bug PR #6904 fixed.

However, moving these tests to the `src` would result in them being included in
distribution packages, which is undesirable. Therefore, we relocate them to the
`.github` directory, ensuring they remain part of the repository but are
excluded from the shipped package.