Transport: set non-interactive default for `--safe-interval` (#6943)
This commit sets the already existing default value for --safe-interval as also non-interactive default.

At the same time, we reduce the value from 30 sec. to 15 seconds. This was long discussed in other issues, see for example #6599 and #6544 there was a debate between 5.0 seconds or 3.0 seconds. 
Either way, those discussion can keep going, meanwhile we reduce here from 30.0 to 15.0 seconds. This should not hurt anybody..