Extend check of stash options when they are specified as empty dict (#6942)
The issue explained in #6945

When using the builder, the stash option is passed as an empty dictionary
without validation. As a result, the stashing task is executed even when no
options arugments are specified, leading to errors when the task tries to access
expected keys in the dictionary.

The hotfix fix extends the check to also verify if the options are not empty.