📚 Transports: improve docstring of `chmod` method
The transport classes have a method for changing the permissions of a file, with
the same name as the UNIX command `chmod`. Users familiar with this command might
not realise the input is an octal number, and that in Python you can express this
as 0o700 to obtain the same outcome as `chmod 700` on the command line. Here we
update the docstring of the `chmod` methods to help clarify this.