devops: Remove flynt hook from pre-commit (#7022)
Most of the rules from the flynt linter are implemented in ruff (FLY category) 
so let's just use that since ruff is much faster.

Interestingly, ruff seems to be better at identifying more places that could be converted into f-strings. 
However, in many of the new cases I felt the current code is more readable
so I just added a noqa comment.