Use TextIOWrapper for opening file from repository in text mode (#6847)
Before this commit, the contents of a file are wrapped in a `StringIO` if a user tries to open a file from the repository in text mode (i.e. with `folderdata.open('somefile.txt', 'r')`). This means the entire file is loaded into memory, which can cause a significant increase in memory usage for large text files. 

This commit, instead wrap the binary stream in a `TextIOWrapper` for streaming based access in text mode to avoid having to load the entire file into memory.

A small change was also necessary for the `cif`-file interface, since apparently the `CifParser` from pymatgen only accepts filenames or `StringIO` streams.