Drop api-auto-docs for now (#7056)
This commit removes API documentation from the main documentation set due to search pollution. We investigated methods to filter these results within RTD; however, we were unable to implement a viable solution at this time. Should a suitable approach be identified, this commit may be reverted. We are proceeding with this merge as an interim solution.