🐛 `TrajectoryData`: Add `pbc`
Currently, the `TrajectoryData` is not functional for structures that are not periodic,
since their `pbc` property is not considered when using the `set_structurelist` method,
which is also used in the constructor.

Here we adapt the class to also accept the `pbc` input in the `set_trajectory` method.
The input should be a list/tuple of length 3 with boolean values, which is validated in
the `_internal_validate` method. It is also optional; in case it is not provided the
current default behavior is maintained, i.e. the structure is assumed to be periodic.
The `pbc` should also be the same for all structures in the trajectory, which is
validated in the `set_structurelist` method. In line with the current approach for
`symbols`, the `pbc` are stored in the attributes and a property is added.

The current changes try to remain backwards-compatible for `get_step_data`, i.e. it now
keeps on returning the same tuple of length 5. However, this does mean that the method
does not return the `pbc`. For the `get_step_structure` method, we intead directly set
the `pbc` from fhe property obtained from the attributes.