Adding ./ in front of a portable code binary (#7080)
This is required if it's in the top folder (of the repository)
and will thus be put in the top folder of the SCRATCH directory
(AiiDA's RemoteData), otherwise the bash script (given to the
scheduler) will not be able to execute it, since `./` is not
typically on the bash's PATH environment variable for security
reasons.

Also, fixing a small bug where the PortableCode will not accept
using a binary defined in a subfolder of the Repository.

Cherry-pick: 9256f2fddd71abecf1e649065ef7dca5c67d061c