Reinstating adaption of verdi process list message
In #7069, we had logic to adapt the output of `verdi process list`
when the process was paused by the handler (and only in this case,
i.e., not when a user was pausing it).

This, in our (C. Pignedoli's and mine) opinion is very important,
otherwise the user most probably is not aware of why the calculation
was paused (or might not even realize it's paused).

To achieve this, a variable is used to mark if the pause was triggered
by the handlers.
Moreover, `on_paused` is overridden to set the process status accordingly.

Importantly, we also changed the exact logic of `restart_and_pause` to
the one we had implemented in the other PR: namely, after pausing,
the `self.ctx.unhandled_failure` is reset to `False`, so that if
another error occurs after replaying, two attempts are tried before
pausing again.