Supervisor for monitoring and keeping alive services
*Motivation*:

For the aiida daemon, a supervisor process for monitoring aiida worker
and keeping them alive, we are using circus. This packages consists of
~15k lines of code and is nowadays mainly maintained by us. It supports
a client server model for supervising processes which is useful for
supervising over a network but not necessary for us. It is also useful
if one wants to go beyond standard UNIX signals to communicate with
processes. Both of these features are not required for the use case
in aiida. I therefore suggest we move to a custom implementation that
is easier to maintain for us.

My second motivation was to implement support for supervising multiple
services with different commands to start. Some of them are incremental
like the aiida worker service some are not. While this could have been
also implemented using circus, the maintenance burden motivated me to
implement this feature in a new custom implementation. This feature is
required to continue using the verdi daemon endpoint for the migration
to airflow. The idea is that verdi daemon is now starting the airflow
scheduler, dag-processor, api-sever and the triggerer. airflow has
for blocking tasks workers that are managed by scheduler. For
nonblocking tasks however airflow does not provide a managing
supervisor. We have to therefore manage the worker for nonblocking tasks
by ourselves. For that use case I want to use this implementation
supporting supervising single process services and worker services
consisting of a tunable amount of processes.

*Implementation*:

Supervisor controller and process:

We need a supervisor process that checks if all services are still
running and if not restarts them. In the current API aiida allows to
modify the number of workers while the supervisor process is already
running. We need to therefore be able to communicate with the supervisor
process the changes of number services (changes of workers). The idea is
that the logic the user can perform through the CLI is in the
ServiceSupervisorController. This one is stateless as the CLI commands
are stateless. The logic for stateful supervisor process that checks if
all process is alive and restarts it if not is in the
ServiceSupervisorProcess.

Furthermore, for better logging separation I organized the supervisor
into sessions. These are folders with timestamps when the daemon
started, so for each daemon restart a new session is created. The logs
for the different processes are also put into separate folder to better
  error messages. One can add utilities to the daemon to merge the logs
  into one file, if this is needed.

```
daemon/
├── profile-<aiida_profile_name>/<session_timestamp>
│   ├── <worker_service_name>/
│   │   ├── worker_service_config.json     # config how to start the worker service
│   │   ├── 1/
│   │   │   ├── info.json          # PID, state, timestamps, failure count
│   │   │   ├── stdout.log         # Service stdout
│   │   │   └── stderr.log         # Service stderr
│   │   └── 2/
│   │       ├── info.json  # PID, state, timestamps, failure count
│   │       ├── stdout.log         # Service stdout
│   │       └── stderr.log         # Service stderr
│   └── <service_name>/
│       ├── service_config.json    # config how to start the service
│       ├── info.json      # PID, state, timestamps, failure count
│       ├── stdout.log             # Service stdout
│       └── stderr.log             # Service stderr
├── supervisor_info.json           # Single daemon PID file
├── supervisor_config.json         # Single daemon PID file
└── supervisor.log                 # Daemon output (background mode only)
```

ServiceConfig:

To add services as a developer one has to inherit from ServiceConfig and
specify the class attribute `service_name` and `command`. The command is
used to start a service process. Each `ServiceConfig` needs to implement
`create_unique_env(self) -> dict[str, str]:` that is the environemnt
variables that should be passed to the process started with the command.
For the aiida worker case it only contains the `AIIDA_PATH`. Because we
need to initialize the ServiceConfig on the ServiceSupervisorController
side to be able to adapt to changes in the supervisor_config.json, we
need some pattern to reconstruct the correct subclass from the service
identifier (the service_name). For this I automatically register
subclasses that set the `service_name`.

*API*:

I only want to expose the AiidaDaemon as part of the public API, while
The supervisor implementation should be internal, but I have not put much
toughts into this yet.

TODO:
- [ ] Cleanup various TODO in the daemon.py
- [ ] Split code in daemon.py into multiple files and subdirectory
- [ ] Remove circus library
- [ ] Integrate fully with CLI (implement missing endpoints and)