👌 `SshTransport`: avoid use of deprecated `getcwd` method
Several methods in the `SshTransport` class still use the deprecated `getcwd` method,
leading to a lot of deprecation warnings being raised when configuring a computer with
`core.ssh`. Although we may decide to deprecate the `core.ssh` transport in favor of
`core.ssh_async` in the future, for now a lot of users are likely still relying on
`core.ssh` and for them the warnings may be confusing/tedious.

Here we adapt the methods of the `SshTransport` class to use `self.sftp.getcwd` instead
of `self.getcwd`. This should change nothing in how `SshTransport` works, since that is
exactly what `self.getcwd` returns as well. However, this avoids correct usage leading
to deprecation warnings and still keeps the deprecation of `getcwd()` in place.