Update error message for SSH transport plugin (#7110)
This commit updates the error message in the SSH transport plugin to provide users with current guidance when SFTP is not supported on the remote computer. The error message now recommends using the core.ssh_async transport plugin with the openssh backend, replacing the outdated reference to the aiida.transports:ssh_only plugin from the deprecated aiida-sshonly package.